MTLAccelerationStructureUsage: u32 => usize

+++ src/acceleration_structure.rs
@@ -9,7 +9,7 @@ use super::*;
 
 bitflags::bitflags! {
     #[derive(Copy, Clone, Debug, Default, Hash, PartialEq, Eq, PartialOrd, Ord)]
-    pub struct MTLAccelerationStructureUsage: u32 {
+    pub struct MTLAccelerationStructureUsage: usize {
         const None = 0;
         const Refit = (1 << 0);
         const PreferFastBuild = (1 << 1);
